refactor: align test file structure with source files
Reorganise test/ so each src/foo.ts has a corresponding test/foo.test.ts:

- Create mock-setup.test.ts (from method-mock + setup-extensions)
- Create mock-expect.test.ts (from expect + expectations-new + method-mock)
- Create mock-spy.test.ts (from spy + serialize + diagnostics)
- Create func.test.ts (extracted from method-mock)
- Move wrap() standalone tests into wrap.test.ts
- Rename stub-class.test.ts → stub.test.ts
- Delete emptied files

No test bodies changed — same 400 tests, same assertions.
