+++ test/mock-spy.test.ts
@@ -1,6 +1,8 @@
 import { describe, it, expect, beforeEach } from 'vitest'
 import deride, { func, stub, wrap, match } from '../src/index'
 
+// ── from spy.test.ts ────────────────────────────────────────────────────
+
 interface Svc {
   greet(name: string): string
   fetch(url: string): Promise<string>
@@ -249,5 +251,189 @@ describe('spy API', () => {
     it('deride.match is exposed', () => {
       expect(deride.match).toBe(match)
     })
+  })
+})
+
+// ── from serialize.test.ts ──────────────────────────────────────────────
+
+interface SerSvc {
+  greet(arg: unknown): unknown
+}
+
+describe('spy.serialize', () => {
+  it('returns the method name and calls', () => {
+    const mock = stub<SerSvc>(['greet'])
+    mock.greet('a')
+    const out = mock.spy.greet.serialize()
+    expect(out.method).toBe('greet')
+    expect(Array.isArray(out.calls)).toBe(true)
+    expect(out.calls.length).toBe(1)
+  })
+
+  it('omits timestamps for snapshot stability', () => {
+    const mock = stub<SerSvc>(['greet'])
+    mock.greet('a')
+    const out = mock.spy.greet.serialize()
+    expect(JSON.stringify(out)).not.toMatch(/"timestamp"/)
+    expect(JSON.stringify(out)).not.toMatch(/"sequence"/)
+  })
+
+  it('sorts object keys for stability', () => {
+    const mock = stub<SerSvc>(['greet'])
+    mock.greet({ b: 2, a: 1, c: 3 })
+    const out = mock.spy.greet.serialize() as { calls: Array<{ args: Array<Record<string, number>> }> }
+    const keys = Object.keys(out.calls[0].args[0])
+    expect(keys).toEqual(['a', 'b', 'c'])
+  })
+
+  it('renders circular refs as [Circular]', () => {
+    const mock = stub<SerSvc>(['greet'])
+    const obj: Record<string, unknown> = { a: 1 }
+    obj.self = obj
+    mock.greet(obj)
+    const out = mock.spy.greet.serialize()
+    expect(JSON.stringify(out)).toContain('[Circular]')
+  })
+
+  it('renders functions as [Function]', () => {
+    const mock = stub<SerSvc>(['greet'])
+    function named() {}
+    mock.greet(named)
+    const out = mock.spy.greet.serialize()
+    expect(JSON.stringify(out)).toContain('[Function: named]')
+  })
+
+  it('matches an inline snapshot for stable usage', () => {
+    const mock = stub<SerSvc>(['greet'])
+    mock.setup.greet.toReturn('hi')
+    mock.greet('alice')
+    mock.greet({ id: 1 })
+    expect(mock.spy.greet.serialize()).toMatchInlineSnapshot(`
+      {
+        "calls": [
+          {
+            "args": [
+              "alice",
+            ],
+            "returned": "hi",
+          },
+          {
+            "args": [
+              {
+                "id": 1,
+              },
+            ],
+            "returned": "hi",
+          },
+        ],
+        "method": "greet",
+      }
+    `)
+  })
+
+  describe('built-in types', () => {
+    it('renders Date as ISO string', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(new Date('2024-01-15T10:30:00Z') as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: string[] }[] }
+      expect(out.calls[0].args[0]).toBe('2024-01-15T10:30:00.000Z')
+    })
+
+    it('renders RegExp as /pattern/flags', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(/foo.*bar/gi as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: string[] }[] }
+      expect(out.calls[0].args[0]).toBe('/foo.*bar/gi')
+    })
+
+    it('renders Error as [Name: message]', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(new TypeError('bad input') as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: string[] }[] }
+      expect(out.calls[0].args[0]).toBe('[TypeError: bad input]')
+    })
+
+    it('renders Map as { __type: "Map", entries: [...] }', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(new Map([['k', 1]]) as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: unknown[] }[] }
+      expect(out.calls[0].args[0]).toEqual({
+        __type: 'Map',
+        entries: [['k', 1]],
+      })
+    })
+
+    it('renders Set as { __type: "Set", values: [...] }', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(new Set([1, 2, 3]) as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: unknown[] }[] }
+      expect(out.calls[0].args[0]).toEqual({
+        __type: 'Set',
+        values: [1, 2, 3],
+      })
+    })
+
+    it('renders Uint8Array as { __type: "Uint8Array", values: [...] }', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(new Uint8Array([10, 20, 30]) as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: unknown[] }[] }
+      expect(out.calls[0].args[0]).toEqual({
+        __type: 'Uint8Array',
+        values: [10, 20, 30],
+      })
+    })
+
+    it('renders bigint and symbol leaves', () => {
+      const sym = Symbol('thing')
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet({ big: 9007199254740993n, sym } as never)
+      const out = mock.spy.greet.serialize() as { calls: { args: { big: string; sym: string }[] }[] }
+      expect(out.calls[0].args[0].big).toBe('9007199254740993n')
+      expect(out.calls[0].args[0].sym).toBe('Symbol(thing)')
+    })
+
+    it('renders nested built-ins recursively (Map of Date)', () => {
+      const mock = stub<SerSvc>(['greet'])
+      mock.greet(new Map([['when', new Date('2024-01-01T00:00:00Z')]]) as never)
+      const out = mock.spy.greet.serialize() as {
+        calls: { args: { __type: string; entries: [string, string][] }[] }[]
+      }
+      expect(out.calls[0].args[0].entries[0][1]).toBe('2024-01-01T00:00:00.000Z')
+    })
+  })
+})
+
+// ── from diagnostics.test.ts (printHistory) ─────────────────────────────
+
+interface PrintSvc {
+  greet(a: unknown, b?: unknown): unknown
+}
+
+describe('printHistory', () => {
+  it('no calls renders the no-calls line', () => {
+    const mock = stub<PrintSvc>(['greet'])
+    expect(mock.spy.greet.printHistory()).toContain('no calls')
+  })
+
+  it('shows method name, args, and return value', () => {
+    const mock = stub<PrintSvc>(['greet'])
+    mock.setup.greet.toReturn('out')
+    mock.greet('in')
+    const out = mock.spy.greet.printHistory()
+    expect(out).toContain('greet')
+    expect(out).toContain("'in'")
+    expect(out).toContain("'out'")
+  })
+
+  it('shows threw -> for throwing calls', () => {
+    const mock = stub<PrintSvc>(['greet'])
+    mock.setup.greet.toThrow('bang')
+    try {
+      mock.greet('x')
+    } catch {
+      /* expected */
+    }
+    const out = mock.spy.greet.printHistory()
+    expect(out).toContain('threw')
   })
 })
